Moved capacitors to FPGA
First I deleted all the Components that were on our board. Then I copy pasted the layout from the Sapristi board. By "Design->Import changes..." all the connections should again be correct. The footprint was thereby changed to the same footprints we had before. Finally I changed to orientation of the capacitors to match with the ones of Sapristi.