Continued Fanout
All pins that won't have to be moved are connected to a via. I made a list of pins which we should move to the edge of the FPGA.
Every capacitor has via on it. 1V5 is always connected to a pin. 3v3 and GND are sometimes not connected to a pin (if there is no pin near it...)